Renamed darling_core::errors to darling_core::error, and made it public
