+++ core/src/lib.rs
@@ -15,7 +15,7 @@ mod macros;
 
 pub mod ast;
 pub mod codegen;
-pub mod errors;
+pub mod error;
 mod from_field;
 mod from_derive_input;
 mod from_meta_item;
@@ -23,7 +23,7 @@ mod from_variant;
 pub mod options;
 pub mod util;
 
-pub use errors::{Result, Error};
+pub use error::{Result, Error};
 pub use from_derive_input::FromDeriveInput;
 pub use from_field::FromField;
 pub use from_meta_item::{FromMetaItem};
